feat: add typescript and fix index and sources
