+++ news-js-main/package.json
@@ -17,16 +17,19 @@
     "license": "ISC",
     "homepage": "",
     "devDependencies": {
+        "@typescript-eslint/eslint-plugin": "^5.59.9",
+        "@typescript-eslint/parser": "^5.59.9",
         "clean-webpack-plugin": "^3.0.0",
         "copy-webpack-plugin": "^7.0.0",
         "css-loader": "^5.1.0",
-        "eslint": "^7.27.0",
+        "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-import": "^2.23.3",
         "eslint-plugin-prettier": "^3.4.0",
         "html-webpack-plugin": "^5.2.0",
         "prettier": "2.2.1",
         "style-loader": "^2.0.0",
+        "typescript": "^5.1.3",
         "webpack": "^5.37.1",
         "webpack-cli": "^4.5.0",
         "webpack-dev-server": "^3.11.2",
