+++ news-js-main/src/components/app/app.ts
