+++ news-js-main/src/components/controller/appLoader.ts
