+++ news-js-main/src/components/controller/controller.ts
