+++ news-js-main/src/components/view/appView.ts
