+++ news-js-main/src/index.ts
@@ -1,5 +1,5 @@
 import App from './components/app/app';
 import './global.css';
 
-const app = new App();
+const app: App = new App();
 app.start();
